Remove failure converter, deprecate constructors that set error causes
