+++ nexus/errors.go
@@ -1,0 +1,301 @@
+package nexus
+
+import (
+	"errors"
+	"fmt"
+)
+
+// An error that directly represents a wire representation of [Failure].
+// The SDK will convert to this error by default unless the [FailureConverter] instance is customized.
+type FailureError struct {
+	// The underlying Failure object this error represents.
+	Failure Failure
+	// An optional nested cause.
+	Cause error
+}
+
+// Error implements the error interface.
+func (e *FailureError) Error() string {
+	return e.Failure.Message
+}
+
+// Unwrap implements the error interface.
+func (e *FailureError) Unwrap() error {
+	return e.Cause
+}
+
+// OperationError represents "failed" and "canceled" operation results.
+type OperationError struct {
+	// Error message.
+	Message string
+	// Stack trace which may be set if this error was generated by a language that supports it.
+	StackTrace string
+	// State of the operation. Only [OperationStateFailed] and [OperationStateCanceled] are valid.
+	State OperationState
+	// The underlying cause for this error.
+	Cause error
+	// Set if this error is constructed from a failure object.
+	OriginalFailure *Failure
+}
+
+// NewOperationFailedError is shorthand for constructing an [OperationError] with state set to [OperationStateFailed]
+// and the given error message as the cause.
+//
+// Deprecated: Use NewOperationFailedErrorf instead.
+func NewOperationFailedError(message string) *OperationError {
+	return &OperationError{
+		State: OperationStateFailed,
+		Cause: errors.New(message),
+	}
+}
+
+// OperationFailedErrorf creates an [OperationError] with state set to [OperationStateFailed], using [fmt.Errorf] to
+// construct the cause.
+//
+// Deprecated: Use NewOperationFailedErrorf instead.
+func OperationFailedErrorf(format string, args ...any) *OperationError {
+	return &OperationError{
+		State: OperationStateFailed,
+		Cause: fmt.Errorf(format, args...),
+	}
+}
+
+// NewOperationFailedErrorf creates an [OperationError] with state set to [OperationStateFailed], using [fmt.Errorf] to
+// construct the message.
+// Errors wrapped via %w formatting are unwrapped and set in the Cause field.
+func NewOperationFailedErrorf(format string, args ...any) *OperationError {
+	err := fmt.Errorf(format, args...)
+	var cause error
+	if len(args) > 0 {
+		cause = errors.Unwrap(err)
+	}
+	return &OperationError{
+		State:   OperationStateFailed,
+		Message: err.Error(),
+		Cause:   cause,
+	}
+}
+
+// NewOperationCanceledError is shorthand for constructing an [OperationError] with state set to
+// [OperationStateCanceled] and the given error message as the cause.
+//
+// Deprecated: Use NewOperationCanceledErrorf instead.
+func NewOperationCanceledError(message string) *OperationError {
+	return &OperationError{
+		State: OperationStateCanceled,
+		Cause: errors.New(message),
+	}
+}
+
+// OperationCanceledErrorf creates an [OperationError] with state set to [OperationStateCanceled], using [fmt.Errorf] to
+// construct the cause.
+//
+// Deprecated: Use NewOperationCanceledErrorf instead.
+func OperationCanceledErrorf(format string, args ...any) *OperationError {
+	message := fmt.Sprintf(format, args...)
+	return &OperationError{
+		State:   OperationStateCanceled,
+		Message: message,
+		// Also setting Cause for compatibility with the old implementation.
+		Cause: errors.New(message),
+	}
+}
+
+// NewOperationCanceledErrorf creates an [OperationError] with state set to [OperationStateCanceled], using [fmt.Errorf]
+// to construct the message.
+// Errors wrapped via %w formatting are unwrapped and set in the Cause field.
+func NewOperationCanceledErrorf(format string, args ...any) *OperationError {
+	err := fmt.Errorf(format, args...)
+	var cause error
+	if len(args) > 0 {
+		cause = errors.Unwrap(err)
+	}
+	return &OperationError{
+		State:   OperationStateCanceled,
+		Message: err.Error(),
+		Cause:   cause,
+	}
+}
+
+// OperationErrorf creates an [OperationError] with the given state, using [fmt.Errorf] to construct the cause.
+//
+// Deprecated: Use NewOperationErrorf instead.
+func OperationErrorf(state OperationState, format string, args ...any) *OperationError {
+	return &OperationError{
+		State: state,
+		Cause: fmt.Errorf(format, args...),
+	}
+}
+
+// NewOperationErrorf creates an [OperationError] with the given state, using [fmt.Errorf] to construct the message.
+// Errors wrapped via %w formatting are unwrapped and set in the Cause field.
+func NewOperationErrorf(state OperationState, format string, args ...any) *OperationError {
+	err := fmt.Errorf(format, args...)
+	var cause error
+	if len(args) > 0 {
+		cause = errors.Unwrap(err)
+	}
+	return &OperationError{
+		State:   OperationStateCanceled,
+		Message: err.Error(),
+		Cause:   cause,
+	}
+}
+
+// Error implements the error interface.
+func (e *OperationError) Error() string {
+	message := fmt.Sprintf("operation %s", e.State)
+	if len(e.Message) > 0 {
+		message += ": " + e.Message
+	}
+	return message
+}
+
+// Unwrap returns the cause for use with utilities in the errors package.
+func (e *OperationError) Unwrap() error {
+	return e.Cause
+}
+
+// HandlerErrorType is an error type associated with a [HandlerError], defined according to the Nexus specification.
+// Only the types defined as consts in this package are valid. Do not use other values.
+type HandlerErrorType string
+
+const (
+	// The server cannot or will not process the request due to an apparent client error. Clients should not retry
+	// this request unless advised otherwise.
+	HandlerErrorTypeBadRequest HandlerErrorType = "BAD_REQUEST"
+	// The client did not supply valid authentication credentials for this request. Clients should not retry
+	// this request unless advised otherwise.
+	HandlerErrorTypeUnauthenticated HandlerErrorType = "UNAUTHENTICATED"
+	// The caller does not have permission to execute the specified operation. Clients should not retry this
+	// request unless advised otherwise.
+	HandlerErrorTypeUnauthorized HandlerErrorType = "UNAUTHORIZED"
+	// The requested resource could not be found but may be available in the future. Clients should not retry
+	// this request unless advised otherwise.
+	HandlerErrorTypeNotFound HandlerErrorType = "NOT_FOUND"
+	// Returned by the server to when it has given up handling a request. The may occur by enforcing a client
+	// provided `Request-Timeout` or for any arbitrary reason such as enforcing some configurable limit. Subsequent
+	// requests by the client are permissible.
+	HandlerErrorTypeRequestTimeout HandlerErrorType = "REQUEST_TIMEOUT"
+	// The request could not be made due to a conflict. The may happen when trying to create an operation that
+	// has already been started. Clients should not retry this request unless advised otherwise.
+	HandlerErrorTypeConflict HandlerErrorType = "CONFLICT"
+	// Some resource has been exhausted, perhaps a per-user quota, or perhaps the entire file system is out of
+	// space. Subsequent requests by the client are permissible.
+	HandlerErrorTypeResourceExhausted HandlerErrorType = "RESOURCE_EXHAUSTED"
+	// An internal error occured. Subsequent requests by the client are permissible.
+	HandlerErrorTypeInternal HandlerErrorType = "INTERNAL"
+	// The server either does not recognize the request method, or it lacks the ability to fulfill the request.
+	// Clients should not retry this request unless advised otherwise.
+	HandlerErrorTypeNotImplemented HandlerErrorType = "NOT_IMPLEMENTED"
+	// The service is currently unavailable. Subsequent requests by the client are permissible.
+	HandlerErrorTypeUnavailable HandlerErrorType = "UNAVAILABLE"
+	// Used by gateways to report that a request to an upstream server has timed out. Subsequent requests by the
+	// client are permissible.
+	HandlerErrorTypeUpstreamTimeout HandlerErrorType = "UPSTREAM_TIMEOUT"
+)
+
+// HandlerErrorRetryBehavior allows handlers to explicity set the retry behavior of a [HandlerError]. If not specified,
+// retry behavior is determined from the error type. For example [HandlerErrorTypeInternal] is not retryable by default
+// unless specified otherwise.
+type HandlerErrorRetryBehavior int
+
+const (
+	// HandlerErrorRetryBehaviorUnspecified indicates that the retry behavior for a [HandlerError] is determined
+	// from the [HandlerErrorType].
+	HandlerErrorRetryBehaviorUnspecified HandlerErrorRetryBehavior = iota
+	// HandlerErrorRetryBehaviorRetryable explicitly indicates that a [HandlerError] should be retried, overriding
+	// the default retry behavior of the [HandlerErrorType].
+	HandlerErrorRetryBehaviorRetryable
+	// HandlerErrorRetryBehaviorNonRetryable explicitly indicates that a [HandlerError] should not be retried,
+	// overriding the default retry behavior of the [HandlerErrorType].
+	HandlerErrorRetryBehaviorNonRetryable
+)
+
+// HandlerError is a special error that can be returned from [Handler] methods for failing a request with a custom
+// status code and failure message.
+type HandlerError struct {
+	// Error Type. Defaults to HandlerErrorTypeInternal.
+	Type HandlerErrorType
+	// Error message.
+	Message string
+	// Stack trace which may be set if this error was generated by a language that supports it.
+	StackTrace string
+	// The underlying cause for this error.
+	Cause error
+	// RetryBehavior of this error. If not specified, retry behavior is determined from the error type.
+	RetryBehavior HandlerErrorRetryBehavior
+	// Set if this error is constructed from a failure object.
+	OriginalFailure *Failure
+}
+
+// HandlerErrorf creates a [HandlerError] with the given type, using [fmt.Errorf] to construct the cause.
+//
+// Deprecated: Use NewHandlerErrorf instead.
+func HandlerErrorf(typ HandlerErrorType, format string, args ...any) *HandlerError {
+	return &HandlerError{
+		Type:  typ,
+		Cause: fmt.Errorf(format, args...),
+	}
+}
+
+// NewHandlerErrorf creates a [HandlerError] with the given type, using [fmt.Errorf] to construct the message.
+// Errors wrapped via %w formatting are unwrapped and set in the Cause field.
+func NewHandlerErrorf(typ HandlerErrorType, format string, args ...any) *HandlerError {
+	err := fmt.Errorf(format, args...)
+	var cause error
+	if len(args) > 0 {
+		cause = errors.Unwrap(err)
+	}
+	return &HandlerError{
+		Type:    typ,
+		Message: err.Error(),
+		Cause:   cause,
+	}
+}
+
+// Retryable returns a boolean indicating whether or not this error is retryable based on the error's RetryBehavior and
+// Type.
+func (e *HandlerError) Retryable() bool {
+	switch e.RetryBehavior {
+	case HandlerErrorRetryBehaviorNonRetryable:
+		return false
+	case HandlerErrorRetryBehaviorRetryable:
+		return true
+	}
+	switch e.Type {
+	case HandlerErrorTypeBadRequest,
+		HandlerErrorTypeUnauthenticated,
+		HandlerErrorTypeUnauthorized,
+		HandlerErrorTypeNotFound,
+		HandlerErrorTypeNotImplemented,
+		HandlerErrorTypeConflict:
+		return false
+	case HandlerErrorTypeResourceExhausted,
+		HandlerErrorTypeInternal,
+		HandlerErrorTypeUnavailable,
+		HandlerErrorTypeUpstreamTimeout,
+		HandlerErrorTypeRequestTimeout:
+		return true
+	default:
+		return true
+	}
+}
+
+// Error implements the error interface.
+func (e *HandlerError) Error() string {
+	typ := e.Type
+	if len(typ) == 0 {
+		typ = HandlerErrorTypeInternal
+	}
+	message := fmt.Sprintf("handler error (%s)", typ)
+	if len(e.Message) > 0 {
+		message += ": " + e.Message
+	}
+	return message
+}
+
+// Unwrap returns the cause for use with utilities in the errors package.
+func (e *HandlerError) Unwrap() error {
+	return e.Cause
+}
